refactor: refactored tests
