+++ tests/pr_test.go
@@ -15,18 +15,11 @@ import (
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testschematic"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/cloudinfo"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testhelper"
 )
 
-const resourceGroup = "geretain-test-base-ocp-vpc"
-const advancedExampleDir = "examples/advanced"
-const basicExampleDir = "examples/basic"
-const fscloudExampleDir = "examples/fscloud"
-const crossKmsSupportExampleDir = "examples/cross_kms_support"
-const customsgExampleDir = "examples/custom_sg"
 const fullyConfigurableTerraformDir = "solutions/fully-configurable"
 
 // Define a struct with fields that match the structure of the YAML data
@@ -34,12 +27,11 @@ const yamlLocation = "../common-dev-assets/common-go-assets/common-permanent-res
 
 // Ensure there is one test per supported OCP version
 const ocpVersion1 = "4.17" // used by TestRunFullyConfigurable, TestRunUpgradeFullyConfigurable, TestFSCloudInSchematic and TestRunMultiClusterExample
-const ocpVersion2 = "4.16" // used by TestCustomSGExample and TestRunCustomsgExample
-const ocpVersion3 = "4.15" // used by TestRunAdvancedExample and TestCrossKmsSupportExample
-const ocpVersion4 = "4.14" // used by TestRunAddRulesToSGExample and TestRunBasicExample
 
-var sharedInfoSvc *cloudinfo.CloudInfoService
-var permanentResources map[string]interface{}
+var (
+	sharedInfoSvc      *cloudinfo.CloudInfoService
+	permanentResources map[string]interface{}
+)
 
 // TestMain will be run before any parallel tests, used to set up a shared InfoService object to track region usage
 // for multiple tests
@@ -55,193 +47,108 @@ func TestMain(m *testing.M) {
 	os.Exit(m.Run())
 }
 
-func setupOptions(t *testing.T, prefix string, terraformDir string, ocpVersion string) *testhelper.TestOptions {
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:          t,
-		TerraformDir:     terraformDir,
-		Prefix:           prefix,
-		ResourceGroup:    resourceGroup,
-		CloudInfoService: sharedInfoSvc,
-		TerraformVars: map[string]interface{}{
-			"ocp_version":     ocpVersion,
-			"access_tags":     permanentResources["accessTags"],
-			"ocp_entitlement": "cloud_pak",
+func validateEnvVariable(t *testing.T, varName string) string {
+	val, present := os.LookupEnv(varName)
+	require.True(t, present, "%s environment variable not set", varName)
+	require.NotEqual(t, "", val, "%s environment variable is empty", varName)
+	return val
+}
+
+func setupTerraform(t *testing.T, prefix, realTerraformDir string) *terraform.Options {
+	tempTerraformDir, err := files.CopyTerraformFolderToTemp(realTerraformDir, prefix)
+	require.NoError(t, err, "Failed to create temporary Terraform folder")
+
+	apiKey := validateEnvVariable(t, "TF_VAR_ibmcloud_api_key")
+	region, err := testhelper.GetBestVpcRegion(apiKey, "../common-dev-assets/common-go-assets/cloudinfo-region-vpc-gen2-prefs.yaml", "eu-de")
+	require.NoError(t, err, "Failed to get best VPC region")
+
+	existingTerraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
+		TerraformDir: tempTerraformDir,
+		Vars: map[string]interface{}{
+			"prefix": prefix,
+			"region": region,
 		},
+		// Set Upgrade to true to ensure latest version of providers and modules are used by terratest.
+		// This is the same as setting the -upgrade=true flag with terraform.
+		Upgrade: true,
 	})
 
-	return options
+	terraform.WorkspaceSelectOrNew(t, existingTerraformOptions, prefix)
+	_, err = terraform.InitAndApplyE(t, existingTerraformOptions)
+	require.NoError(t, err, "Init and Apply of temp existing resource failed")
+
+	return existingTerraformOptions
+}
+
+func cleanupTerraform(t *testing.T, options *terraform.Options, prefix string) {
+	if t.Failed() && strings.ToLower(os.Getenv("DO_NOT_DESTROY_ON_FAILURE")) == "true" {
+		fmt.Println("Terratest failed. Debug the test and delete resources manually.")
+		return
+	}
+	logger.Log(t, "START: Destroy (existing resources)")
+	terraform.Destroy(t, options)
+	terraform.WorkspaceDelete(t, options, prefix)
+	logger.Log(t, "END: Destroy (existing resources)")
 }
 
 func TestRunFullyConfigurableInSchematics(t *testing.T) {
 	t.Parallel()
 
-	// ------------------------------------------------------------------------------------
 	// Provision resources first
-	// ------------------------------------------------------------------------------------
 	prefix := fmt.Sprintf("ocp-fc-%s", strings.ToLower(random.UniqueId()))
-	realTerraformDir := "./existing-resources"
-	tempTerraformDir, _ := files.CopyTerraformFolderToTemp(realTerraformDir, prefix)
+	existingTerraformOptions := setupTerraform(t, prefix, "./existing-resources")
 
-	// Verify ibmcloud_api_key variable is set
-	checkVariable := "TF_VAR_ibmcloud_api_key"
-	val, present := os.LookupEnv(checkVariable)
-	require.True(t, present, checkVariable+" environment variable not set")
-	require.NotEqual(t, "", val, checkVariable+" environment variable is empty")
-	region, _ := testhelper.GetBestVpcRegion(val, "../common-dev-assets/common-go-assets/cloudinfo-region-vpc-gen2-prefs.yaml", "eu-de")
-
-	logger.Log(t, "Tempdir: ", tempTerraformDir)
-	existingTerraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
-		TerraformDir: tempTerraformDir,
-		Vars: map[string]interface{}{
-			"prefix": prefix,
-			"region": region,
-		},
-		// Set Upgrade to true to ensure latest version of providers and modules are used by terratest.
-		// This is the same as setting the -upgrade=true flag with terraform.
-		Upgrade: true,
+	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
+		Testing:                t,
+		Prefix:                 "ocp-fc",
+		TarIncludePatterns:     []string{"*.tf", fullyConfigurableTerraformDir + "/*.*", "scripts/*.sh", "kubeconfig/README.md"},
+		TemplateFolder:         fullyConfigurableTerraformDir,
+		Tags:                   []string{"test-schematic"},
+		DeleteWorkspaceOnFail:  false,
+		WaitJobCompleteMinutes: 60,
 	})
 
-	terraform.WorkspaceSelectOrNew(t, existingTerraformOptions, prefix)
-	_, existErr := terraform.InitAndApplyE(t, existingTerraformOptions)
-	if existErr != nil {
-		assert.True(t, existErr == nil, "Init and Apply of temp existing resource failed")
-	} else {
-
-		options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
-			Testing: t,
-			Prefix:  "ocp-fc",
-			TarIncludePatterns: []string{
-				"*.tf",
-				fullyConfigurableTerraformDir + "/*.*",
-				"scripts/*.sh",
-				"kubeconfig/README.md",
-			},
-			TemplateFolder:         fullyConfigurableTerraformDir,
-			Tags:                   []string{"test-schematic"},
-			DeleteWorkspaceOnFail:  false,
-			WaitJobCompleteMinutes: 60,
-		})
-
-		options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-			{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-			{Name: "prefix", Value: options.Prefix, DataType: "string"},
-			{Name: "cluster_name", Value: "cluster", DataType: "string"},
-			{Name: "ocp_version", Value: ocpVersion1, DataType: "string"},
-			{Name: "ocp_entitlement", Value: "cloud_pak", DataType: "string"},
-			{Name: "existing_resource_group_name", Value: terraform.Output(t, existingTerraformOptions, "resource_group_name"), DataType: "string"},
-			{Name: "existing_cos_instance_crn", Value: terraform.Output(t, existingTerraformOptions, "cos_instance_id"), DataType: "string"},
-			{Name: "existing_vpc_crn", Value: terraform.Output(t, existingTerraformOptions, "vpc_crn"), DataType: "string"},
-		}
-
-		err := options.RunSchematicTest()
-		assert.Nil(t, err, "This should not have errored")
-
+	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
+		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
+		{Name: "prefix", Value: options.Prefix, DataType: "string"},
+		{Name: "cluster_name", Value: "cluster", DataType: "string"},
+		{Name: "ocp_version", Value: ocpVersion1, DataType: "string"},
+		{Name: "ocp_entitlement", Value: "cloud_pak", DataType: "string"},
+		{Name: "existing_resource_group_name", Value: terraform.Output(t, existingTerraformOptions, "resource_group_name"), DataType: "string"},
+		{Name: "existing_cos_instance_crn", Value: terraform.Output(t, existingTerraformOptions, "cos_instance_id"), DataType: "string"},
+		{Name: "existing_vpc_crn", Value: terraform.Output(t, existingTerraformOptions, "vpc_crn"), DataType: "string"},
 	}
-
-	envVal, _ := os.LookupEnv("DO_NOT_DESTROY_ON_FAILURE")
-	// Destroy the temporary existing resources if required
-	if t.Failed() && strings.ToLower(envVal) == "true" {
-		fmt.Println("Terratest failed. Debug the test and delete resources manually.")
-	} else {
-		logger.Log(t, "START: Destroy (existing resources)")
-		terraform.Destroy(t, existingTerraformOptions)
-		terraform.WorkspaceDelete(t, existingTerraformOptions, prefix)
-		logger.Log(t, "END: Destroy (existing resources)")
-	}
-}
-
-func getClusterIngress(options *testhelper.TestOptions) error {
-
-	// Get output of the last apply
-	outputs, outputErr := terraform.OutputAllE(options.Testing, options.TerraformOptions)
-	if !assert.NoError(options.Testing, outputErr, "error getting last terraform apply outputs: %s", outputErr) {
-		return nil
-	}
-
-	// Validate that the "cluster_name" key is present in the outputs
-	expectedOutputs := []string{"cluster_name"}
-	_, ValidationErr := testhelper.ValidateTerraformOutputs(outputs, expectedOutputs...)
-
-	// Proceed with the cluster ingress health check if "cluster_name" is valid
-	if assert.NoErrorf(options.Testing, ValidationErr, "Some outputs not found or nil: %s", ValidationErr) {
-		options.CheckClusterIngressHealthyDefaultTimeout(outputs["cluster_name"].(string))
-	}
-	return nil
+	require.NoError(t, options.RunSchematicTest(), "This should not have errored")
+	cleanupTerraform(t, existingTerraformOptions, prefix)
 }
 
 func TestRunUpgradeFullyConfigurable(t *testing.T) {
 	t.Parallel()
 
-	// ------------------------------------------------------------------------------------
 	// Provision existing resources first
-	// ------------------------------------------------------------------------------------
 	prefix := fmt.Sprintf("ocp-existing-%s", strings.ToLower(random.UniqueId()))
-	realTerraformDir := "./existing-resources"
-	tempTerraformDir, _ := files.CopyTerraformFolderToTemp(realTerraformDir, fmt.Sprintf(prefix+"-%s", strings.ToLower(random.UniqueId())))
+	existingTerraformOptions := setupTerraform(t, prefix, "./existing-resources")
 
-	// Verify ibmcloud_api_key variable is set
-	checkVariable := "TF_VAR_ibmcloud_api_key"
-	val, present := os.LookupEnv(checkVariable)
-	require.True(t, present, checkVariable+" environment variable not set")
-	require.NotEqual(t, "", val, checkVariable+" environment variable is empty")
-	region, _ := testhelper.GetBestVpcRegion(val, "../common-dev-assets/common-go-assets/cloudinfo-region-vpc-gen2-prefs.yaml", "eu-de")
-
-	logger.Log(t, "Tempdir: ", tempTerraformDir)
-	existingTerraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
-		TerraformDir: tempTerraformDir,
-		Vars: map[string]interface{}{
-			"prefix": prefix,
-			"region": region,
-		},
-		// Set Upgrade to true to ensure latest version of providers and modules are used by terratest.
-		// This is the same as setting the -upgrade=true flag with terraform.
-		Upgrade: true,
+	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
+		Testing:                t,
+		Prefix:                 "ocp-fc-upg",
+		TarIncludePatterns:     []string{"*.tf", "scripts/*.sh", "kubeconfig/README.md", fullyConfigurableTerraformDir + "/*.*"},
+		TemplateFolder:         fullyConfigurableTerraformDir,
+		Tags:                   []string{"test-schematic"},
+		DeleteWorkspaceOnFail:  false,
+		WaitJobCompleteMinutes: 60,
 	})
 
-	terraform.WorkspaceSelectOrNew(t, existingTerraformOptions, prefix)
-	_, existErr := terraform.InitAndApplyE(t, existingTerraformOptions)
-	if existErr != nil {
-		assert.True(t, existErr == nil, "Init and Apply of temp existing resource failed")
-	} else {
-
-		options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
-			Testing: t,
-			Prefix:  "ocp-fc-upg",
-			TarIncludePatterns: []string{
-				"*.tf",
-				"scripts/*.sh",
-				"kubeconfig/README.md",
-				fullyConfigurableTerraformDir + "/*.*",
-			},
-			TemplateFolder:         fullyConfigurableTerraformDir,
-			Tags:                   []string{"test-schematic"},
-			DeleteWorkspaceOnFail:  false,
-			WaitJobCompleteMinutes: 60,
-		})
-
-		options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-			{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-			{Name: "prefix", Value: options.Prefix, DataType: "string"},
-			{Name: "cluster_name", Value: "cluster", DataType: "string"},
-			{Name: "ocp_version", Value: ocpVersion1, DataType: "string"},
-			{Name: "existing_resource_group_name", Value: terraform.Output(t, existingTerraformOptions, "resource_group_name"), DataType: "string"},
-			{Name: "existing_cos_instance_crn", Value: terraform.Output(t, existingTerraformOptions, "cos_instance_id"), DataType: "string"},
-			{Name: "existing_vpc_crn", Value: terraform.Output(t, existingTerraformOptions, "vpc_crn"), DataType: "string"},
-		}
-
-		err := options.RunSchematicUpgradeTest()
-		assert.Nil(t, err, "This should not have errored")
-
+	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
+		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
+		{Name: "prefix", Value: options.Prefix, DataType: "string"},
+		{Name: "cluster_name", Value: "cluster", DataType: "string"},
+		{Name: "ocp_version", Value: ocpVersion1, DataType: "string"},
+		{Name: "existing_resource_group_name", Value: terraform.Output(t, existingTerraformOptions, "resource_group_name"), DataType: "string"},
+		{Name: "existing_cos_instance_crn", Value: terraform.Output(t, existingTerraformOptions, "cos_instance_id"), DataType: "string"},
+		{Name: "existing_vpc_crn", Value: terraform.Output(t, existingTerraformOptions, "vpc_crn"), DataType: "string"},
 	}
 
-	envVal, _ := os.LookupEnv("DO_NOT_DESTROY_ON_FAILURE")
-	// Destroy the temporary existing resources if required
-	if t.Failed() && strings.ToLower(envVal) == "true" {
-		fmt.Println("Terratest failed. Debug the test and delete resources manually.")
-	} else {
-		logger.Log(t, "START: Destroy (existing resources)")
-		terraform.Destroy(t, existingTerraformOptions)
-		terraform.WorkspaceDelete(t, existingTerraformOptions, prefix)
-		logger.Log(t, "END: Destroy (existing resources)")
-	}
+	require.NoError(t, options.RunSchematicUpgradeTest(), "This should not have errored")
+	cleanupTerraform(t, existingTerraformOptions, prefix)
 }
